feat: ringkas sekali pak
